+++ src/handlers/__init__.py
@@ -1,0 +1,4 @@
+from src.handlers.echo_message import echo_message
+from src.handlers.welcome import send_welcome
+
+__all__ = ['send_welcome', 'echo_message']
