+++ src/handlers/echo_message.py
@@ -1,0 +1,6 @@
+from main import bot
+
+# Handle all other messages with content_type 'text' (content_types defaults to ['text'])
+@bot.message_handler(func=lambda message: True)
+async def echo_message(message):
+    await bot.reply_to(message, message.text)
