+++ src/handlers/welcome.py
@@ -1,0 +1,7 @@
+from main import bot
+
+# Handle '/start' and '/help'
+@bot.message_handler(commands=['help', 'start'])
+async def send_welcome(message):
+    text = 'Hi, I am EchoBot.\nJust write me something and I will repeat it!'
+    await bot.reply_to(message, text)
